git-svn-id: file:///var/www/svn2/oecdo/oecde@717 50b15a66-aedb-45cf-97b2-84293224aabc

+++ discussion/src/main/java/org/sgu/oecde/discussion/dao/IRootDao.java
@@ -19,7 +19,7 @@ public interface IRootDao extends IUpdateDao<Root>{
      * @return количество постов
      * @throws DataAccessException
      */
-    public int getNodesCount(int idObject, ForumTypes typeObject)  throws DataAccessException ;
+    public int getNodesCount(Long idObject, ForumTypes typeObject)  throws DataAccessException ;
 
     /**
      * сохраняет данный корневой элемент
